refactor: abilityConfig -> turnAbilityConfig
