+++ src/lib/gpu/injection/generateWgsl.ts
@@ -6,7 +6,7 @@ import { indent } from 'lib/gpu/injection/wgslUtils'
 import { GpuConstants, RelicsByPart } from 'lib/gpu/webgpuTypes'
 import computeShader from 'lib/gpu/wgsl/computeShader.wgsl?raw'
 import structs from 'lib/gpu/wgsl/structs.wgsl?raw'
-import { AbilityKind, getAbilityKind } from 'lib/optimization/rotation/abilityConfig'
+import { AbilityKind, getAbilityKind } from 'lib/optimization/rotation/turnAbilityConfig'
 import { SortOption } from 'lib/optimization/sortOptions'
 import { Form } from 'types/form'
 import { OptimizerContext } from 'types/optimizer'
