+++ src/lib/optimization/rotation/rotationPreprocessor.ts
@@ -1,7 +1,7 @@
-import { toTurnAbility } from 'lib/optimization/rotation/abilityConfig'
 import { CastoricePreprocessor } from 'lib/optimization/rotation/preprocessor/preprocessCharacters'
 import { ScholarLostInEruditionPreprocessor, WavestriderCaptainPreprocessor } from 'lib/optimization/rotation/preprocessor/preprocessSets'
 import { AbilityPreprocessorBase } from 'lib/optimization/rotation/preprocessor/preprocessUtils'
+import { toTurnAbility } from 'lib/optimization/rotation/turnAbilityConfig'
 import { preprocessTurnAbilities } from 'lib/optimization/rotation/turnPreprocessor'
 import { ComboState } from 'lib/tabs/tabOptimizer/combo/comboDrawerController'
 import { Form } from 'types/form'
