+++ src/types/form.d.ts
@@ -1,5 +1,5 @@
 import { SetsOrnaments, SetsRelics } from 'lib/constants/constants'
-import { AbilityKind, TurnAbilityName } from 'lib/optimization/rotation/abilityConfig'
+import { AbilityKind, TurnAbilityName } from 'lib/optimization/rotation/turnAbilityConfig'
 import { SortOption } from 'lib/optimization/sortOptions'
 import { Simulation, SimulationRequest } from 'lib/simulations/statSimulationTypes'
 import { SetConditionals } from 'lib/tabs/tabOptimizer/combo/comboDrawerController'
