+++ src/types/optimizer.d.ts
@@ -3,7 +3,7 @@ import { ElementName, PathName } from 'lib/constants/constants'
 import { DynamicConditional } from 'lib/gpu/conditionals/dynamicConditionals'
 import { ConditionalRegistry } from 'lib/optimization/calculateConditionals'
 import { ComputedStatsArray } from 'lib/optimization/computedStatsArray'
-import { AbilityKind } from 'lib/optimization/rotation/abilityConfig'
+import { AbilityKind } from 'lib/optimization/rotation/turnAbilityConfig'
 import { CharacterConditionalsController, ConditionalValueMap, LightConeConditionalsController } from 'types/conditionals'
 import { ElementalDamageType, ElementalResPenType } from 'types/metadata'
